refactor: 뷰 파일 구조 개선 및 네이밍 변경
