+++ src/js/view/index.js
@@ -1,0 +1,6 @@
+export { default as ProductTab } from './ProductTab/ProductTab';
+export { default as ChargeTab } from './ChargeTab/ChargeTab';
+export { default as PurchaseTab } from './PurchaseTab/PurchaseTab';
+export { default as LoginPage } from './LoginPage/LoginPage';
+export { default as RegisterPage } from './RegisterPage/RegisterPage';
+export { default as UserInfoPage } from './UserInfoPage/UserInfoPage';
